+++ sample/Main.hs
@@ -4,10 +4,29 @@
 module Main (main) where
 
 import GHC.IO.Encoding (setLocaleEncoding, utf8)
+import Plutarch (
+    Config (Config, tracingMode),
+    TracingMode (NoTracing),
+    compile,
+ )
 import Plutarch.Api.V1 (PValidator)
 import Plutarch.Extra.TermCont (ptryFromC)
-import Plutarch.Prelude (PAsData, PInteger, Term, pfromData, pif, plam, popaque, ptraceError, unTermCont, (#==))
-import Plutarch.Test.Precompiled (fromPTerm, (@!>), (@&), (@>))
+import Plutarch.Prelude (
+    PAsData,
+    PInteger,
+    Term,
+    pdata,
+    pfromData,
+    pif,
+    plam,
+    popaque,
+    ptraceError,
+    unTermCont,
+    (#),
+    (#==),
+    (:-->),
+ )
+import Plutarch.Test.Precompiled (fromPTerm, testEqualityCase, (@!>), (@&), (@>))
 import Test.Tasty (TestTree, defaultMain, testGroup)
 
 import qualified PlutusTx as PlutusTx
@@ -19,6 +38,7 @@ main = do
         testGroup
             "test suite"
             [ sampleValidatorTest
+            , sampleFunctionTest
             ]
 
 sampleValidator :: Term s PValidator
@@ -35,9 +55,29 @@ sampleValidatorTest :: TestTree
 sampleValidatorTest = fromPTerm "sample validator" sampleValidator $ do
     [PlutusTx.toData (), PlutusTx.toData (1 :: Integer), PlutusTx.toData ()] @> "It should succeed when given 1"
     [PlutusTx.toData (), PlutusTx.toData (10 :: Integer), PlutusTx.toData ()] @!> "It should fail when given 10"
-    [PlutusTx.toData (), PlutusTx.toData (), PlutusTx.toData ()] @> "It should fail when given non integer"
+    [PlutusTx.toData (), PlutusTx.toData (), PlutusTx.toData ()] @!> "It should fail when given non integer"
 
     [PlutusTx.toData ()] @& do
         [PlutusTx.toData (1 :: Integer), PlutusTx.toData ()] @> "(Sharing first argument) It should succeed when given 1"
         [PlutusTx.toData (10 :: Integer), PlutusTx.toData ()] @!> "(Sharing first argument) It should fail when given 10"
         [PlutusTx.toData (), PlutusTx.toData ()] @!> "(Sharing first argument) It should fail when given non integer"
+
+sampleFunction :: Term s (PAsData PInteger :--> PAsData PInteger :--> PAsData PInteger)
+sampleFunction = plam $ \x y -> pdata ((pfromData x) + (pfromData y) + 1)
+
+sampleFunctionTest :: TestTree
+sampleFunctionTest = fromPTerm "sample function" sampleFunction $ do
+    [PlutusTx.toData (1 :: Integer), PlutusTx.toData (1 :: Integer)] @> "It should not fail"
+
+    testEqualityCase "1" compiled1 [PlutusTx.toData (1 :: Integer)]
+    testEqualityCase "1 1" compiled11 [PlutusTx.toData (1 :: Integer), PlutusTx.toData (1 :: Integer)]
+  where
+    compiled1 =
+        either (error . show) id $
+            compile (Config{tracingMode = NoTracing}) $
+                sampleFunction # (pdata 1)
+
+    compiled11 =
+        either (error . show) id $
+            compile (Config{tracingMode = NoTracing}) $
+                sampleFunction # (pdata 1) # (pdata 1)