+++ src/Plutarch/Test/Precompiled.hs
@@ -18,6 +18,7 @@ module Plutarch.Test.Precompiled (
     testEvalCase,
     (@>),
     (@!>),
+    testEqualityCase,
     fromPTerm,
 ) where
 
@@ -61,11 +62,17 @@ data Expectation
 
  @since 0.1.1.0
 -}
-data TestCase = TestCase
-    { dScript :: DebuggableScript
-    , caseName :: String
-    , expectation :: Expectation
-    }
+data TestCase
+    = EvalTestCase
+        { dScript :: DebuggableScript
+        , caseName :: String
+        , expectation :: Expectation
+        }
+    | EqualityTestCase
+        { dScript :: DebuggableScript
+        , expectedScript :: Script
+        , caseName :: String
+        }
 
 ourStyle :: Style
 ourStyle = style{lineLength = 80}
@@ -75,9 +82,12 @@ ppLogs = \case
     [] -> "No logs found. Did you forget to build with +development?"
     logs -> vcat $ ppDoc <$> logs
 
+failWithStyle :: Doc -> Result
+failWithStyle = testFailed . renderStyle ourStyle
+
 instance IsTest TestCase where
     testOptions = Tagged []
-    run _ TestCase{..} _ = return $
+    run _ EvalTestCase{..} _ = return $
         case (r, expectation) of
             (Right _, Success) -> testPassed ""
             (Right x, Failure) -> failWithStyle . unexpectedSuccess x $ dt
@@ -86,8 +96,6 @@ instance IsTest TestCase where
       where
         (r, _, _) = evalScript $ script dScript
         (_, _, dt) = evalScript $ debugScript dScript
-        failWithStyle :: Doc -> Result
-        failWithStyle = testFailed . renderStyle ourStyle
         unexpectedFailure :: EvalError -> [Text] -> Doc
         unexpectedFailure err logs =
             "Expected a successful run, but failed instead.\n"
@@ -96,6 +104,23 @@ instance IsTest TestCase where
         unexpectedSuccess :: Script -> [Text] -> Doc
         unexpectedSuccess result logs =
             "Expected a failing run, but succeeded instead.\n"
+                <> hang "Result" 4 (ppDoc result)
+                <> hang "Logs" 4 (ppLogs logs)
+    run _ EqualityTestCase{..} _ = return $
+        case (r, e) of
+            (Right x, Right y) ->
+                if x == y
+                    then testPassed ""
+                    else testFailed "Two script does not match"
+            (Left err, _) -> failWithStyle . failedToEvaluate err $ dt
+            (_, Left err) -> failWithStyle . failedToEvaluate err $ mempty
+      where
+        (r, _, _) = evalScript $ script dScript
+        (_, _, dt) = evalScript $ debugScript dScript
+        (e, _, _) = evalScript $ expectedScript
+        failedToEvaluate :: EvalError -> [Text] -> Doc
+        failedToEvaluate result logs =
+            "Script evaluation failed, both scripts need to suceed in order to check equality.\n"
                 <> hang "Result" 4 (ppDoc result)
                 <> hang "Logs" 4 (ppLogs logs)
 
@@ -130,6 +155,23 @@ withApplied args tests = local (flip applyDebuggableScript args) tests
 (@&) :: [Data] -> TestCompiled () -> TestCompiled ()
 args @& tests = withApplied args tests
 
+{- | Tests if script is eqaul to expectation or not given arguments.
+
+ @since 0.1.1.0
+-}
+testEqualityCase :: String -> Script -> [Data] -> TestCompiled ()
+testEqualityCase name e args = do
+    ds <- ask
+    let applied = applyDebuggableScript ds args
+        testCase =
+            EqualityTestCase
+                { dScript = applied
+                , expectedScript = e
+                , caseName = name
+                }
+    tell $ pure testCase
+    return ()
+
 {- | Tests if script succeed or not given arguments.
 
  @since 0.1.1.0
@@ -139,7 +181,7 @@ testEvalCase name e args = do
     ds <- ask
     let applied = applyDebuggableScript ds args
         testCase =
-            TestCase
+            EvalTestCase
                 { dScript = applied
                 , caseName = name
                 , expectation = e